# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/deposit:
    post:
      operationId: deposit
      tags: [payments]
      summary: Deposit funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "200":
          description: Deposit completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
  /payments/withdraw:
    post:
      operationId: withdraw
      tags: [payments]
      summary: Withdraw funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "200":
          description: Withdrawal completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      properties:
        amount:
          type: integer
      required:
        - amount
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: integer
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - amount
        - type
        - status
        - createdAt
...